backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.69",
    "fastapi[cors]>=0.121.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]
